Update signature of memmove to also take 32 bit parameter

+++ system/include/libc/string.h
@@ -23,7 +23,7 @@ extern "C" {
 #include <bits/alltypes.h>
 
 void *memcpy (void *__restrict, const void *__restrict, unsigned int);
-void *memmove (void *, const void *, size_t);
+void *memmove (void *, const void *, unsigned int);
 //LLVM expects param3 of memset (which is an llvm intrinsic) to correspond to the size of an llvm pointer (4bytes)
 //We had added modifications that make the C pointer 8 bytes. So change the sig of memset
 // void *memset (void *, int, size_t);
